+++ src/lib/src/attribution-dialog/attribution-dialog.service.ts
@@ -5,7 +5,7 @@ import { MdDialog, MdDialogRef, MdDialogConfig } from '@angular/material';
 import { AttributionDialogComponent } from './attribution-dialog.component';
 import { MimeResizeService } from './../core/mime-resize-service/mime-resize.service';
 import { AttributionDialogResizeService } from './attribution-dialog-resize.service';
-import { MimeDomHelper } from './../core/mime-dom-renderer';
+import { MimeDomHelper } from './../core/mime-dom-helper';
 import { Dimensions } from './../core/models/dimensions';
 
 @Injectable()
@@ -87,10 +87,10 @@ export class AttributionDialogService {
   private getPosition(el: ElementRef) {
     const padding = 20;
     const dimensions = new MimeDomHelper().getBoundingClientRect(el);
-    return {
+    return new Dimensions({
       top: dimensions.top + dimensions.height - this.attributionDialogHeight - padding,
       left: dimensions.left + padding
-    };
+    });
   }
 
 }