+++ src/lib/src/contents-dialog/contents-dialog-config-strategy.ts
@@ -1,4 +1,4 @@
-import { MimeDomHelper } from './../core/mime-dom-renderer';
+import { MimeDomHelper } from './../core/mime-dom-helper';
 import { ElementRef } from '@angular/core';
 import { MdDialogConfig } from '@angular/material';
 
@@ -39,10 +39,10 @@ export class DesktopContentsDialogConfigStrategy implements ContentsDialogConfig
 
   private getPosition(el: ElementRef): Dimensions {
     const dimensions = new MimeDomHelper().getBoundingClientRect(el);
-    return {
+    return new Dimensions({
       top: dimensions.top + 64,
       left: dimensions.right - DesktopContentsDialogConfigStrategy.dialogWidth - DesktopContentsDialogConfigStrategy.paddingRight
-    };
+    });
   }
 
 }