+++ src/lib/src/viewer/viewer-header/viewer-header.component.ts
@@ -7,6 +7,8 @@ import { MimeViewerIntl } from './../../core/viewer-intl';
 import { Manifest } from './../../core/models/manifest';
 import { ContentsDialogComponent } from './../../contents-dialog/contents-dialog.component';
 import { ContentsDialogService } from './../../contents-dialog/contents-dialog.service';
+import { MimeDomHelper } from '../../core/mime-dom-helper';
+import { FullscreenService } from './../../core/fullscreen-service/fullscreen.service';
 
 @Component({
   selector: 'mime-viewer-header',
@@ -16,14 +18,23 @@ import { ContentsDialogService } from './../../contents-dialog/contents-dialog.s
 })
 export class ViewerHeaderComponent implements OnInit, OnDestroy {
   private subscriptions: Array<Subscription> = [];
+  isFullscreenEnabled = false;
 
   constructor(
     public intl: MimeViewerIntl,
     private changeDetectorRef: ChangeDetectorRef,
-    private contentsDialogService: ContentsDialogService) { }
+    private contentsDialogService: ContentsDialogService,
+    private fullscreenService: FullscreenService) { }
 
   ngOnInit() {
+    this.isFullscreenEnabled = this.fullscreenService.isEnabled();
+
     this.subscriptions.push(this.intl.changes.subscribe(() => this.changeDetectorRef.markForCheck()));
+
+    this.subscriptions.push(this.fullscreenService.onChange.subscribe(() => {
+      this.changeDetectorRef.detectChanges();
+    }));
+
   }
 
   ngOnDestroy() {
@@ -34,6 +45,14 @@ export class ViewerHeaderComponent implements OnInit, OnDestroy {
 
   public openContents() {
     this.contentsDialogService.toggle();
+  }
+
+  public toggleFullscreen(): void {
+    return new MimeDomHelper().toggleFullscreen();
+  }
+
+  public isInFullScreen(): boolean {
+    return this.fullscreenService.isFullscreen();
   }
 
 }